feat: SC.2 — per-grid voxel scale, CPU cross-grid shadows
A cross-grid shadow has two scale factors, handled on the two sides of
the world-space occlusion test:

- Caster (grid being shaded): WorldShadowCtx gains voxel_world_size;
  WorldShadow::occluded lifts the grid-local voxel shadow ray to world by
  scaling the local position AND direction by vws (max_t is a caster-voxel
  distance, so the world segment covers t·vws)
- Occluder (each grid tested): GridOcc gains vws; occluded_in_grid divides
  the inverse-rotated world ray (o and d) by vws to reach that grid's voxel
  frame — dividing both preserves the ray parameter t, so max_t still clips
  correctly

A grid shadowing itself is scale-invariant (the two factors cancel) and
vws==1.0 everywhere is byte-identical to XS.1. shadow_max_dist is a
caster-voxel cap (world reach = shadow_max_dist·vws), consistent with
BakeLight.radius / StreamRadius.

Test sc2_scaled_grid_casts_world_correct_shadow: a vws=2 occluder block
filling the same world box as an unscaled block casts a matching shadow
(within ~30%, not just darker); negative-verified — without the
occluder-side /vws the world ray misses the voxel AABB and the shadow
drops to zero. Dense test oracle scaled to match. 210 scene tests green,
vws=1.0 byte-identical.

+++ crates/roxlap-core/src/dda_sprite.rs
@@ -394,7 +394,7 @@ impl SpriteOccluder {
 }
 
 impl WorldOccluder for SpriteOccluder {
-    fn occluded_world(&self, origin: [f32; 3], dir: [f32; 3], max_t: f32) -> bool {
+    fn occluded_world(&self, origin: [f64; 3], dir: [f32; 3], max_t: f32) -> bool {
         self.entries
             .iter()
             .any(|e| sprite_entry_occluded(e, origin, dir, max_t))
@@ -409,9 +409,17 @@ fn bb_scale3(v: [f32; 3], k: f32) -> [f32; 3] {
     [v[0] * k, v[1] * k, v[2] * k]
 }
 
-fn sprite_entry_occluded(e: &SpriteOccEntry, ow: [f32; 3], dw: [f32; 3], max_t: f32) -> bool {
-    // World → sprite-local voxel space (same transform as the draw).
-    let rel = [ow[0] - e.pos[0], ow[1] - e.pos[1], ow[2] - e.pos[2]];
+fn sprite_entry_occluded(e: &SpriteOccEntry, ow: [f64; 3], dw: [f32; 3], max_t: f32) -> bool {
+    // World → sprite-local voxel space (same transform as the draw). SC.2 —
+    // the origin subtraction runs in f64 (both are large world coords, so this
+    // avoids catastrophic cancellation) then narrows to f32 for the local
+    // basis apply; the sprite frame's coords are small.
+    #[allow(clippy::cast_possible_truncation)]
+    let rel = [
+        (ow[0] - f64::from(e.pos[0])) as f32,
+        (ow[1] - f64::from(e.pos[1])) as f32,
+        (ow[2] - f64::from(e.pos[2])) as f32,
+    ];
     let ol = mat_apply(&e.minv, rel);
     let origin = [ol[0] + e.pivot[0], ol[1] + e.pivot[1], ol[2] + e.pivot[2]];
     let dir = mat_apply(&e.minv, dw);
@@ -2313,7 +2321,7 @@ mod tests {
             "vws=2 @ unit basis == vws=1 @ 2x basis, pixel for pixel"
         );
 
-        let occludes = |dense: &SpriteDense, basis: f32, x: f32| {
+        let occludes = |dense: &SpriteDense, basis: f32, x: f64| {
             let mut occ = SpriteOccluder::new();
             occ.push(
                 Arc::new(dense.clone()),
